Extract ProjectConfig
