+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/pages/PartIdentification.tsx
@@ -1,482 +1,17 @@
-import React, { useState, useEffect, useRef, useCallback, useMemo } from 'react';
-import {
-  Divider,
-  Text,
-  Flex,
-  Dropdown,
-  DropdownItemProps,
-  Checkbox,
-  Input,
-  Alert,
-  ShorthandCollection,
-} from '@fluentui/react-northstar';
-import { useHistory } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import Axios from 'axios';
-import * as R from 'ramda';
-
-import { thunkGetProject, thunkPostProject, updateProjectData } from '../store/project/projectActions';
-import { Project, ProjectData } from '../store/project/projectTypes';
-import { State } from '../store/State';
-import { formatDropdownValue, Value } from '../util/formatDropdownValue';
-import { getAppInsights } from '../TelemetryService';
-import { WarningDialog } from '../components/WarningDialog';
-import { AddCameraLink } from '../components/AddModuleDialog/AddCameraLink';
-import { AddLocationLink } from '../components/AddModuleDialog/AddLocationLink';
-import { AddPartLink } from '../components/AddModuleDialog/AddPartLink';
-import { LabelImage } from '../store/image/imageTypes';
-import { getLabelImages } from '../store/image/imageActions';
-import { Button } from '../components/Button';
-
-const sendTrainInfoToAppInsight = async (selectedParts): Promise<void> => {
-  const { data: images } = await Axios.get('/api/images/');
-
-  const selectedPartIds = selectedParts.map((e) => e.id);
-  const interestedImagesLength = images.filter((e) => selectedPartIds.includes(e.part)).length;
-  const appInsight = getAppInsights();
-  if (appInsight)
-    appInsight.trackEvent({
-      name: 'train',
-      properties: {
-        images: interestedImagesLength,
-        parts: selectedParts.length,
-        source: '',
-      },
-    });
-};
+import React from 'react';
+import { Flex } from '@fluentui/react-northstar';
+import { ProjectConfig } from '../components/ProjectConfigPanel/ProjectConfig';
+import { LiveViewContainer } from '../components/LiveViewContainer';
 
 export const PartIdentification: React.FC = () => {
-  const dispatch = useDispatch();
-  const { isLoading, error, data } = useSelector<State, Project>((state) => state.project);
-  const {
-    id: projectId,
-    camera,
-    location,
-    parts,
-    needRetraining,
-    accuracyRangeMin,
-    accuracyRangeMax,
-    maxImages: maxImage,
-    sendMessageToCloud,
-    framesPerMin,
-    accuracyThreshold,
-  } = data;
-  const allImages = useSelector<State, LabelImage[]>((state) => state.images);
-  const images = useMemo(() => allImages.filter((e) => !e.is_relabel), [allImages]);
-  const [isTestModel, setIsTestModel] = useState(false);
-  const [cameraLoading, dropDownCameras, selectedCamera, setSelectedCameraById] = useDropdownItems<any>(
-    'cameras',
-    isTestModel,
-  );
-  const [partLoading, dropDownParts, selectedParts, setSelectedPartsById] = useDropdownItems<any>(
-    'parts',
-    isTestModel,
-    true,
-  );
-  const [locationLoading, dropDownLocations, selectedLocations, setSelectedLocationById] = useDropdownItems<
-    any
-  >('locations', isTestModel);
-  const history = useHistory();
-  const [maxImgCountError, setMaxImgCountError] = useState(false);
-  const [suggestMessage, setSuggestMessage] = useState({ min: 0, max: 0, partName: '', rangeMessage: '' });
-  const hasUserUpdateAccuracyRange = useRef(false);
-
-  useEffect(() => {
-    if (!cameraLoading && !partLoading && !locationLoading) {
-      dispatch(thunkGetProject(isTestModel));
-    }
-  }, [dispatch, cameraLoading, locationLoading, partLoading, isTestModel]);
-
-  useEffect(() => {
-    dispatch(getLabelImages());
-  }, [dispatch]);
-
-  useEffect(() => {
-    if (!isTestModel) {
-      if (location) setSelectedLocationById(location);
-      if (parts.length) setSelectedPartsById(parts);
-      if (camera) setSelectedCameraById(camera);
-    }
-  }, [
-    camera,
-    isTestModel,
-    location,
-    parts,
-    setSelectedCameraById,
-    setSelectedLocationById,
-    setSelectedPartsById,
-  ]);
-
-  const handleSubmitConfigure = async (): Promise<void> => {
-    try {
-      if (!isTestModel) sendTrainInfoToAppInsight(selectedParts);
-
-      const id = await dispatch(
-        thunkPostProject(projectId, selectedLocations, selectedParts, selectedCamera, isTestModel),
-      );
-
-      if (typeof id !== 'undefined')
-        history.push(`/cameras/detail?name=${selectedCamera.name}&isDemo=${isTestModel}`);
-    } catch (e) {
-      alert(e);
-    }
-  };
-
-  const setData = (keyName: keyof ProjectData, value: ProjectData[keyof ProjectData]): void => {
-    dispatch(updateProjectData({ [keyName]: value }));
-  };
-
-  useEffect(() => {
-    const partsWithImageLength = images.reduce((acc, cur) => {
-      const { id } = cur.part;
-      const relatedPartIdx = acc.findIndex((e) => e.id === id);
-      if (relatedPartIdx >= 0) acc[relatedPartIdx].length = acc[relatedPartIdx].length + 1 || 1;
-      return acc;
-    }, R.clone(selectedParts));
-
-    const minimumLengthPart = partsWithImageLength.reduce(
-      (acc, cur) => {
-        if (cur.length < acc.length) return { name: cur.name, length: cur.length };
-        return acc;
-      },
-      { name: '', length: Infinity },
-    );
-
-    if (minimumLengthPart.length === Infinity) return;
-    if (minimumLengthPart.length < 30) {
-      if (!hasUserUpdateAccuracyRange.current)
-        dispatch(updateProjectData({ accuracyRangeMax: 40, accuracyRangeMin: 10 }));
-      setSuggestMessage({
-        min: 10,
-        max: 40,
-        partName: minimumLengthPart.name,
-        rangeMessage: 'lower than 30',
-      });
-    } else if (minimumLengthPart.length >= 30 && minimumLengthPart.length < 80) {
-      if (!hasUserUpdateAccuracyRange.current)
-        dispatch(updateProjectData({ accuracyRangeMax: 60, accuracyRangeMin: 30 }));
-      setSuggestMessage({
-        min: 30,
-        max: 60,
-        partName: minimumLengthPart.name,
-        rangeMessage: 'between 30 to 80',
-      });
-    } else if (minimumLengthPart.length >= 80 && minimumLengthPart.length < 130) {
-      if (!hasUserUpdateAccuracyRange.current)
-        dispatch(updateProjectData({ accuracyRangeMax: 80, accuracyRangeMin: 50 }));
-      setSuggestMessage({
-        min: 50,
-        max: 80,
-        partName: minimumLengthPart.name,
-        rangeMessage: 'between 80 to 130',
-      });
-    } else if (minimumLengthPart.length >= 130) {
-      if (!hasUserUpdateAccuracyRange.current)
-        dispatch(updateProjectData({ accuracyRangeMax: 90, accuracyRangeMin: 60 }));
-      setSuggestMessage({
-        min: 60,
-        max: 90,
-        partName: minimumLengthPart.name,
-        rangeMessage: 'more than 130',
-      });
-    }
-  }, [accuracyRangeMin, dispatch, images, selectedParts]);
-
-  const accracyRangeDisabled = !needRetraining || isTestModel;
-  const messageToCloudDisabled = !sendMessageToCloud;
-
   return (
-    <Flex hAlign="center" styles={{ width: '600px' }} column gap="gap.medium">
-      <Text size="larger" weight="semibold">
-        Part Identification
-      </Text>
-      <Divider color="black" styles={{ width: '100%' }} />
-      {error && (
-        <Alert danger header="Load Part Identification Error" content={`${error.name}: ${error.message}`} />
-      )}
-      <Flex column>
-        <ModuleSelector
-          moduleName="camera"
-          to="/cameras"
-          value={selectedCamera}
-          setSelectedModuleItem={setSelectedCameraById}
-          items={dropDownCameras}
-          isMultiple={false}
-        />
-        <ModuleSelector
-          moduleName="parts"
-          to="/parts"
-          value={selectedParts}
-          setSelectedModuleItem={setSelectedPartsById}
-          items={dropDownParts}
-          isMultiple={true}
-        />
-        <ModuleSelector
-          moduleName="location"
-          to="/locations"
-          value={selectedLocations}
-          setSelectedModuleItem={setSelectedLocationById}
-          items={dropDownLocations}
-          isMultiple={false}
-          isTestModel={isTestModel}
-        />
-      </Flex>
-      <Flex styles={{ height: '250px' }}>
-        <Flex column gap="gap.medium" styles={{ width: '50%' }}>
-          <Checkbox
-            label="Set up retraining"
-            checked={needRetraining}
-            onChange={(_, { checked }): void => setData('needRetraining', checked)}
-            disabled={isTestModel}
-          />
-          <Text disabled={accracyRangeDisabled} weight="bold">
-            Accuracy for capture image
-          </Text>
-          <Flex gap="gap.small">
-            <Flex column hAlign="center">
-              <Text disabled={accracyRangeDisabled} content="Minimum: " />
-              <Input
-                type="number"
-                disabled={accracyRangeDisabled}
-                inline
-                value={accuracyRangeMin}
-                onChange={(_, { value }): void => {
-                  if (!hasUserUpdateAccuracyRange.current) hasUserUpdateAccuracyRange.current = true;
-                  setData('accuracyRangeMin', value);
-                }}
-                icon="%"
-                fluid
-              />
-            </Flex>
-            <Flex column hAlign="center">
-              <Text disabled={accracyRangeDisabled}>Maximum: </Text>
-              <Input
-                type="number"
-                disabled={accracyRangeDisabled}
-                inline
-                value={accuracyRangeMax}
-                onChange={(_, { value }): void => {
-                  if (!hasUserUpdateAccuracyRange.current) hasUserUpdateAccuracyRange.current = true;
-                  setData('accuracyRangeMax', value);
-                }}
-                icon="%"
-                fluid
-              />
-            </Flex>
-          </Flex>
-          <Text styles={{ fontSize: '12px' }} success>
-            {`The Part ${suggestMessage.partName} contains images ${suggestMessage.rangeMessage}, recommend to set the range to Min ${suggestMessage.min}% and Max ${suggestMessage.max}% `}
-          </Text>
-          <Flex column hAlign="center">
-            <Text disabled={accracyRangeDisabled}>Maximum Images to Store: </Text>
-            <Input
-              type="number"
-              disabled={accracyRangeDisabled}
-              inline
-              value={maxImage}
-              onChange={(_, { value }): void => {
-                if ((value as any) < 15) setMaxImgCountError(true);
-                else setMaxImgCountError(false);
-                setData('maxImages', value);
-              }}
-            />
-            {maxImgCountError && <Text error>Cannot be less than 15</Text>}
-          </Flex>
-        </Flex>
-        <Divider vertical color="black" styles={{ margin: '0px 10px' }} />
-        <Flex column gap="gap.medium">
-          <Checkbox
-            label="Send message to cloud"
-            checked={sendMessageToCloud}
-            onChange={(_, { checked }): void => setData('sendMessageToCloud', checked)}
-          />
-          <Flex column hAlign="center">
-            <Text disabled={messageToCloudDisabled}>Frames per minute: </Text>
-            <Input
-              type="number"
-              disabled={messageToCloudDisabled}
-              inline
-              value={framesPerMin}
-              onChange={(_, { value }): void => setData('framesPerMin', value)}
-            />
-          </Flex>
-          <Flex column hAlign="center">
-            <Text disabled={messageToCloudDisabled}>Accuracy threshold: </Text>
-            <Input
-              type="number"
-              disabled={messageToCloudDisabled}
-              inline
-              value={accuracyThreshold}
-              onChange={(_, { value }): void => setData('accuracyThreshold', value)}
-            />
-          </Flex>
-        </Flex>
-      </Flex>
-      <Button
-        content="Configure"
-        primary
-        onClick={handleSubmitConfigure}
-        disabled={(!selectedCamera || !selectedLocations || !selectedParts || isLoading) && !isTestModel}
-        loading={isLoading}
-        circular
-      />
-      <Divider color="black" styles={{ width: '100%' }} />
-      <Text>Try pretrained detection</Text>
-      <TestModelButton isTestModel={isTestModel} setIsTestModel={setIsTestModel} />
-    </Flex>
-  );
-};
-
-const TestModelButton = ({ isTestModel, setIsTestModel }): JSX.Element => {
-  if (isTestModel) {
-    return <Button content="Back" onClick={(): void => setIsTestModel(false)} primary />;
-  }
-
-  return (
-    <WarningDialog
-      confirmButton="Confirm"
-      onConfirm={(): void => setIsTestModel(true)}
-      contentText={
-        <>
-          <p>
-            &quot;Demo Pretrained Detection&quot; is for seeing inference result, no retraining experience
-            here.
-          </p>
-          <p>For retraining experience, please create a new model</p>
-        </>
-      }
-      trigger={<Button content="Pretrained Detection" primary circular />}
-    />
-  );
-};
-
-// TODO Make this integrate with Redux
-function useDropdownItems<T>(
-  moduleName: string,
-  isTestModel: boolean,
-  isMultiple?: boolean,
-): [boolean, DropdownItemProps[], T | T[], (id: string | string[]) => void] {
-  const originItems = useRef<(T & { id: number })[]>([]);
-  const [dropDownItems, setDropDownItems] = useState<DropdownItemProps[]>([]);
-  const [selectedItem, setSelectedItem] = useState<T | T[]>(isMultiple ? [] : null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const url =
-      moduleName === 'cameras' && isTestModel
-        ? `/api/${moduleName}/`
-        : `/api/${moduleName}/?is_demo=${Number(isTestModel)}`;
-    setLoading(true);
-    Axios(url)
-      .then(({ data }) => {
-        if (data.length === 0) {
-          setDropDownItems([
-            {
-              header: `No ${moduleName.replace('s', '')} found, please add ${moduleName.replace('s', '')}`,
-              content: {
-                key: 'Dummy',
-              },
-            },
-          ]);
-        } else {
-          setDropDownItems(
-            data.map((e) => ({
-              header: e.name,
-              content: {
-                key: e.id,
-              },
-            })),
-          );
-        }
-        originItems.current = data;
-        if (isMultiple) {
-          setSelectedItem(data);
-        } else {
-          setSelectedItem(data[0]);
-        }
-        setLoading(false);
-        return void 0;
-      })
-      .catch((err) => console.error(err))
-      .finally(() => setLoading(false));
-  }, [isMultiple, moduleName, isTestModel]);
-
-  const setSelectedItemById = useCallback((id: string | string[]): void => {
-    if (Array.isArray(id)) {
-      const correspondedItems = id.reduce((acc, cur) => {
-        const correspondedItem = originItems.current.find((ele) => ele.id.toString(10) === cur.toString());
-        if (correspondedItem) acc.push(correspondedItem);
-        return acc;
-      }, []);
-      setSelectedItem(correspondedItems as any);
-    } else {
-      const correspondedItem = originItems.current.find((ele) => ele.id.toString(10) === id.toString());
-      if (correspondedItem) setSelectedItem(correspondedItem);
-    }
-  }, []);
-
-  return [loading, dropDownItems, selectedItem, setSelectedItemById];
-}
-
-/* Module Selector */
-
-type ModuleSelectorProps = {
-  moduleName: string;
-  to: string;
-  value: Value;
-  setSelectedModuleItem: (id: string | string[]) => void;
-  items: ShorthandCollection<DropdownItemProps>;
-  isMultiple: boolean;
-  isTestModel?: boolean;
-};
-
-const ModuleSelector: React.FC<ModuleSelectorProps> = ({
-  moduleName,
-  to,
-  value,
-  setSelectedModuleItem,
-  items,
-  isMultiple,
-  isTestModel,
-}): JSX.Element => {
-  const onDropdownChange = (_, data): void => {
-    if (data.value === null) return;
-    if (Array.isArray(data.value)) {
-      const ids = data.value.map((ele) => ele.content.key);
-      setSelectedModuleItem(ids);
-    } else {
-      const { key } = data.value.content;
-      setSelectedModuleItem(key);
-    }
-  };
-
-  const getAddModuleLink = (): JSX.Element => {
-    switch (moduleName) {
-      case 'camera':
-        return <AddCameraLink />;
-      case 'location':
-        return <AddLocationLink />;
-      case 'parts':
-        return <AddPartLink />;
-      default:
-        return <p>Un supported module</p>;
-    }
-  };
-
-  return (
-    <Flex vAlign="center" gap="gap.medium">
-      <Text styles={{ width: '90px' }}>{`Select ${moduleName}`}</Text>
-      {isTestModel ? (
-        <Dropdown items={items} value={formatDropdownValue(value)} multiple={isMultiple} open={false} />
-      ) : (
-        <Dropdown
-          items={items}
-          onChange={onDropdownChange}
-          value={formatDropdownValue(value)}
-          multiple={isMultiple}
-        />
-      )}
-      {getAddModuleLink()}
+    <Flex>
+      <div style={{ flexGrow: 2 }}>
+        <ProjectConfig />
+      </div>
+      <div style={{ flexGrow: 3 }}>
+        <LiveViewContainer showVideo={true} initialAOIData={{ useAOI: false, AOIs: [] }} cameraId={0} />
+      </div>
     </Flex>
   );
 };
